Release DataGenerator 1.1, clean Repo, Readme done
