+++ DataGenerator/src/main/java/dataGenerator/LoadFileController.java
@@ -1,4 +1,4 @@
-package DataProgram;
+package dataGenerator;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
