+++ DataGenerator/src/main/java/dataGenerator/SceneController.java
@@ -1,4 +1,4 @@
-package DataProgram;
+package dataGenerator;
 
 import javafx.scene.Parent;
 import javafx.scene.Scene;
