+++ DataGenerator/src/main/resources/views/JudgmentRequested.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.VBox?>
 
 <AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.172-ea"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="DataProgram.JudgmentRequestedController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="dataGenerator.JudgmentRequestedController">
     <VBox prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
           AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
         <Label text="Judge the messages below : Yes if it's a good message, No if it's a bad message."/>
