+++ DataGenerator/src/main/resources/views/MessageComponent.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="57.0" prefWidth="498.0" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="DataProgram.MessageComponentController">
+<AnchorPane prefHeight="57.0" prefWidth="498.0" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="dataGenerator.MessageComponentController">
     <fx:define>
         <ToggleGroup fx:id="YesNo" />
     </fx:define>
