--- src/DataProgram/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name = 'DataProgram'
-
